# docs/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCare Hub API
  version: 1.0.0
  description: >
    API para gerenciamento de clientes, veículos, serviços, peças, ordens de
    serviço, orçamentos, aprovações e acompanhamento de status de uma oficina
    mecânica.
servers:
  - url: http://localhost:8080
    description: Ambiente local
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/customers/{customerId}/service-orders:
    get:
      tags:
        - Service Orders
      summary: Lista ordens de serviço de um cliente
      operationId: listServiceOrdersByCustomer
      parameters:
        - $ref: '#/components/parameters/CustomerIdPathParam'
      responses:
        '200':
          description: Ordens de serviço encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/service-orders:
    post:
      tags:
        - Service Orders
      summary: Cria uma ordem de serviço identificando o cliente por CPF ou CNPJ
      description: >
        Abre uma Ordem de Serviço recebendo dados do cliente, veículo, serviços e peças/insumos.
        Retorna o identificador único da OS no campo `id`.
      operationId: createServiceOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceOrderRequest'
            examples:
              clienteVeiculoExistentes:
                summary: OS para cliente e veículo já cadastrados
                description: Use este exemplo depois de autenticar como admin. Ele já gera orçamento automaticamente.
                value:
                  customerDocument: '12345678909'
                  vehicleId: 20000000-0000-0000-0000-000000000001
                  diagnosticNotes: Cliente relata ruido ao frear e vibração no pedal.
                  services:
                    - serviceId: 30000000-0000-0000-0000-000000000004
                      quantity: 1
                  parts:
                    - partId: 40000000-0000-0000-0000-000000000005
                      quantity: 1
                    - partId: 40000000-0000-0000-0000-000000000007
                      quantity: 1
                  generateBudget: true
              clienteVeiculoNovos:
                summary: OS cadastrando cliente e veículo no mesmo fluxo
                description: Demonstra identificação por CPF/CNPJ, cadastro de veículo e geração automática de orçamento.
                value:
                  customerDocument: '52998224725'
                  customer:
                    name: Yasmin Barcelos
                    phone: '11987654321'
                    email: yasmin.cliente@example.com
                    address:
                      street: Avenida Paulista
                      number: '1000'
                      complement: Sala 10
                      neighborhood: Bela Vista
                      city: São Paulo
                      state: SP
                      zipCode: 01310-100
                  vehicle:
                    plate: TST1A23
                    brand: Honda
                    model: Civic Touring
                    year: 2021
                    mileage: 42000
                  diagnosticNotes: Cliente solicita revisão preventiva antes de viagem.
                  services:
                    - serviceId: 30000000-0000-0000-0000-000000000003
                      quantity: 1
                  parts:
                    - partId: 40000000-0000-0000-0000-000000000001
                      quantity: 4
                    - partId: 40000000-0000-0000-0000-000000000002
                      quantity: 1
                  generateBudget: true
      responses:
        '201':
          description: Ordem de serviço criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
              example:
                id: 50000000-0000-0000-0000-000000000010
                customerId: 10000000-0000-0000-0000-000000000001
                vehicleId: 20000000-0000-0000-0000-000000000001
                status: WAITING_APPROVAL
                diagnosticNotes: Cliente relata ruido ao frear e vibração no pedal.
                services:
                  - serviceId: 30000000-0000-0000-0000-000000000004
                    name: Diagnóstico de freios
                    quantity: 1
                    unitPrice: 180.0
                    totalPrice: 180.0
                parts:
                  - partId: 40000000-0000-0000-0000-000000000005
                    name: Pastilha de freio dianteira
                    quantity: 1
                    unitPrice: 220.0
                    totalPrice: 220.0
                servicesTotal: 180.0
                partsTotal: 220.0
                totalAmount: 400.0
                createdAt: '2026-07-05T10:00:00-03:00'
                budgetGeneratedAt: '2026-07-05T10:01:00-03:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
        - Service Orders
      summary: Lista ordens de serviço
      description: >
        Lista a fila operacional de OS. Sem filtro de status, a listagem da Fase 2 retorna apenas OS abertas
        ordenadas por prioridade/status: IN_PROGRESS, WAITING_APPROVAL, IN_DIAGNOSIS e RECEIVED.
        Dentro do mesmo status, as mais antigas aparecem primeiro; FINISHED e DELIVERED não aparecem na fila.
      operationId: listServiceOrders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ServiceOrderStatus'
        - name: customerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: vehicleId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Ordens de serviço encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderListResponse'
              example:
                items:
                  - id: 50000000-0000-0000-0000-000000000020
                    customerId: 10000000-0000-0000-0000-000000000001
                    vehicleId: 20000000-0000-0000-0000-000000000001
                    status: IN_PROGRESS
                    diagnosticNotes: Serviço em execução.
                    services: [ ]
                    parts: [ ]
                    servicesTotal: 0.0
                    partsTotal: 0.0
                    totalAmount: 0.0
                    createdAt: '2026-07-05T08:00:00-03:00'
                  - id: 50000000-0000-0000-0000-000000000021
                    customerId: 10000000-0000-0000-0000-000000000002
                    vehicleId: 20000000-0000-0000-0000-000000000002
                    status: WAITING_APPROVAL
                    diagnosticNotes: Aguardando aprovação do orçamento.
                    services: [ ]
                    parts: [ ]
                    servicesTotal: 0.0
                    partsTotal: 0.0
                    totalAmount: 0.0
                    createdAt: '2026-07-05T09:00:00-03:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/service-orders/tracking:
    get:
      tags:
        - Service Orders
      summary: Acompanha o progresso de ordens de serviço pelo cliente
      description: Consulte por id da OS ou por CPF/CNPJ do cliente junto com a placa do veículo.
      operationId: trackServiceOrders
      parameters:
        - name: serviceOrderId
          in: query
          required: false
          example: 50000000-0000-0000-0000-000000000002
          schema:
            type: string
            format: uuid
        - name: customerDocument
          in: query
          required: false
          example: '12345678909'
          schema:
            type: string
            minLength: 11
            maxLength: 18
            pattern: "^\\d{11}$|^\\d{14}$|^\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}$|^\\d{2}\\.\\d{3}\\.\\d{3}/\\d{4}-\\d{2}$"
        - name: plate
          in: query
          required: false
          example: MCA1D23
          schema:
            type: string
            minLength: 7
            maxLength: 8
            pattern: '^[A-Za-z]{3}-?[0-9]{4}$|^[A-Za-z]{3}-?[0-9][A-Za-z][0-9]{2}$'
      responses:
        '200':
          description: Progresso da ordem encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderTrackingListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}:
    get:
      tags:
        - Service Orders
      summary: Busca uma ordem de serviço por id
      description: Consulta os dados atuais da OS, incluindo o status atual.
      operationId: getServiceOrderById
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      responses:
        '200':
          description: Ordem de serviço encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
              example:
                id: 50000000-0000-0000-0000-000000000002
                customerId: 10000000-0000-0000-0000-000000000001
                vehicleId: 20000000-0000-0000-0000-000000000001
                status: WAITING_APPROVAL
                diagnosticNotes: Cliente solicita revisão de freios.
                services: [ ]
                parts: [ ]
                servicesTotal: 0.0
                partsTotal: 0.0
                totalAmount: 0.0
                createdAt: '2026-07-05T10:00:00-03:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/services:
    post:
      tags:
        - Service Orders
      summary: Adiciona um serviço a uma ordem de serviço
      operationId: addServiceToServiceOrder
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddServiceOrderServiceRequest'
            examples:
              adicionarServico:
                summary: Inclui serviço solicitado na OS
                value:
                  serviceId: 30000000-0000-0000-0000-000000000005
                  quantity: 1
      responses:
        '200':
          description: Serviço adicionado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/parts:
    post:
      tags:
        - Service Orders
      summary: Adiciona uma peça a uma ordem de serviço
      operationId: addPartToServiceOrder
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddServiceOrderPartRequest'
            examples:
              adicionarPeca:
                summary: Inclui peça ou insumo na OS
                value:
                  partId: 40000000-0000-0000-0000-000000000005
                  quantity: 1
      responses:
        '200':
          description: Peça adicionada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/budget/generate:
    post:
      tags:
        - Service Orders
      summary: Gera o orçamento de uma ordem de serviço
      operationId: generateServiceOrderBudget
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      responses:
        '200':
          description: Orçamento gerado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/budget/approve:
    post:
      tags:
        - Service Orders
      summary: Aprova o orçamento de uma ordem de serviço
      description: Aprova o orçamento existente e move a OS para execução quando as regras de negócio permitirem.
      operationId: approveServiceOrderBudget
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
      responses:
        '200':
          description: Orçamento aprovado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
              example:
                id: 50000000-0000-0000-0000-000000000002
                customerId: 10000000-0000-0000-0000-000000000001
                vehicleId: 20000000-0000-0000-0000-000000000001
                status: WAITING_APPROVAL
                diagnosticNotes: Cliente solicita revisão de freios.
                services: [ ]
                parts: [ ]
                servicesTotal: 0.0
                partsTotal: 0.0
                totalAmount: 0.0
                createdAt: '2026-07-05T10:00:00-03:00'
                approvedAt: '2026-07-05T10:10:00-03:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/budget/decision:
    post:
      tags:
        - Service Orders
      summary: Recebe decisão externa de aprovação ou recusa do orçamento
      description: >
        Endpoint demonstrável para simular uma notificação externa, como email, aprovando ou recusando o orçamento.
        Não exige login de usuário. Exige o header X-External-Service-Token com o valor configurado em EXTERNAL_SERVICE_TOKEN.
        Quando aprovado, confirma a reserva de peças. Quando recusado, libera reservas e retorna a OS para diagnóstico.
      operationId: decideServiceOrderBudget
      security: [ ]
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
        - $ref: '#/components/parameters/ExternalServiceTokenHeaderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalBudgetDecisionRequest'
            examples:
              aprovarPorEmail:
                summary: Aprovação simulada por email
                value:
                  decision: APPROVED
                  source: email
                  reason: Cliente aprovou o orçamento pelo link de notificação.
              recusarPorEmail:
                summary: Recusa simulada por email
                value:
                  decision: REJECTED
                  source: email
                  reason: Cliente recusou o orçamento e solicitou revisão.
      responses:
        '200':
          description: Decisão registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
              examples:
                aprovado:
                  summary: Orçamento aprovado por notificação externa
                  value:
                    id: 50000000-0000-0000-0000-000000000002
                    customerId: 10000000-0000-0000-0000-000000000001
                    vehicleId: 20000000-0000-0000-0000-000000000001
                    status: WAITING_APPROVAL
                    diagnosticNotes: Cliente solicita revisão de freios.
                    services: [ ]
                    parts: [ ]
                    servicesTotal: 0.0
                    partsTotal: 0.0
                    totalAmount: 0.0
                    createdAt: '2026-07-05T10:00:00-03:00'
                    approvedAt: '2026-07-05T10:12:00-03:00'
                recusado:
                  summary: Orçamento recusado por notificação externa
                  value:
                    id: 50000000-0000-0000-0000-000000000002
                    customerId: 10000000-0000-0000-0000-000000000001
                    vehicleId: 20000000-0000-0000-0000-000000000001
                    status: IN_DIAGNOSIS
                    diagnosticNotes: Cliente solicita revisão de freios.
                    services: [ ]
                    parts: [ ]
                    servicesTotal: 0.0
                    partsTotal: 0.0
                    totalAmount: 0.0
                    createdAt: '2026-07-05T10:00:00-03:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/service-orders/{serviceOrderId}/budget/external-approval:
    post:
      tags:
        - Service Orders
      summary: Recebe aprovação externa do orçamento
      description: >
        Webhook demonstrável para registrar aprovação de orçamento recebida por uma ferramenta externa,
        como email. Não exige login de usuário. Exige o header X-External-Service-Token e usa as mesmas regras
        de domínio do fluxo principal de aprovação.
      operationId: approveServiceOrderBudgetFromExternalTool
      security: [ ]
      parameters:
        - $ref: '#/components/parameters/ServiceOrderIdPathParam'
        - $ref: '#/components/parameters/ExternalServiceTokenHeaderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalBudgetNotificationRequest'
            examples:
              aprovarPorEmail:
                summary: Aprovação recebida por email
                value:
                  source: email
                  reason: Cliente aprovou o orçamento pelo link de notificação.
      responses:
        '200':
          description: Aprovação externa registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderResponse'
              example:
                id: 50000000-0000-0000-0000-000000000002
                customerId: 10000000-0000-0000-0000-000000000001
                vehicleId: 20000000-0000-0000-0000-000000000001
                status: WAITING_APPROVAL
                diagnosticNotes: Cliente solicita revisão de freios.
                services: [ ]
                parts: [ ]
                servicesTotal: 0.0
                partsTotal: 0.0
                totalAmount: 0.0
                createdAt: '2026-07-05T10:00:00-03:00'
                approvedAt: '2026-07-05T10:12:00-03:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    CustomerIdPathParam:
      name: customerId
      in: path
      required: true
      example: 10000000-0000-0000-0000-000000000001
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requisição invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 400
            error: Bad Request
            message: Invalid request body
            path: /api/v1/service-orders
            details:
              - customerDocument: must not be blank
    Unauthorized:
      description: Credenciais ausentes ou invalidas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 401
            error: Unauthorized
            message: Missing or invalid credentials
            path: /api/v1/customers
            details: [ ]
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-06-20T17:00:00-03:00'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/v1/customers/00000000-0000-0000-0000-000000000000
            details: [ ]
...
  schemas:
...
    ServiceOrderListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ServiceOrderResponse'
...